python-dist/requirements.txt
argon2-cffi==23.1.0
email_validator==2.1.1
fastapi==0.115.11
loguru==0.7.2
pydantic==2.6.3
pydantic-settings==2.2.1
pyjwt==2.8.0
tortoise-orm==0.21.6
uvicorn==0.27.1
passlib==1.7.4
boto3==1.34.58
python-multipart==0.0.18
starlette==0.40.0
setuptools==68.2.2
mangum==0.17.0
aiomysql==0.2.0
asyncpg==0.29.0
aiosqlite==0.17.0
requests==2.31.0
Pillow==10.2.0
watchdog==3.0.0
aiohttp==3.9.1